Server/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
colorama==0.4.4
hexdump==3.3
httpie==2.4.0
idna==2.10
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
more-itertools==8.7.0
Naked==0.1.31
portend==2.7.1
pyasn1==0.4.8
pycryptodome==3.10.1
Pygments==2.8.0
PySocks==1.7.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
shellescape==3.8.1
six==1.15.0
tempora==4.0.1
urllib3==1.26.3
zc.lockfile==2.0
